Inflection-47 Integrate en Wikidata into Unicode Inflection
