+++ inflection/resources/org/unicode/inflection/dictionary/.gitattributes
@@ -1,2 +1,4 @@
 dictionary_da.lst filter=lfs diff=lfs merge=lfs -text
+dictionary_en.lst filter=lfs diff=lfs merge=lfs -text
 inflectional_da.xml filter=lfs diff=lfs merge=lfs -text
+inflectional_en.xml filter=lfs diff=lfs merge=lfs -text
