+++ inflection/tools/dictionary-parser/src/main/resources/org/unicode/wikidata/filter_en.properties
@@ -1,0 +1,18 @@
+# Copyright 2025 Unicode Incorporated and others. All rights reserved.
+#
+# These are lexemes that should either be ignored due to irrelevance that can't be easily tagged as irrelevant,
+# or words that are just not that common that should be sorted last in the inflection patterns.
+L15388=rare
+L299075=omit
+L342586=omit
+L468896=omit
+L469033=omit
+L469036=omit
+L469037=omit
+L469040=omit
+L469047=omit
+L684798=omit
+L685028=omit
+L685030=omit
+L984169=omit
+L1321935=omit
